*: convert to uint64 (#2668)
Convert `Duty.Slot` from int64 to uint64.

This is recommended according to ethereum consensus specs: https://github.com/ethereum/consensus-specs/blob/dev/specs/phase0/beacon-chain.md#custom-types

category: feature 
ticket: #2658

+++ core/consensus/roundtimer.go
@@ -32,7 +32,7 @@ func getTimerFunc() timerFunc {
 		}
 
 		return func(duty core.Duty) roundTimer {
-			random := rand.New(rand.NewSource(int64(duty.Type) + duty.Slot)) //nolint:gosec // Required for consistent pseudo-randomness.
+			random := rand.New(rand.NewSource(int64(uint64(duty.Type) + duty.Slot))) //nolint:gosec // Required for consistent pseudo-randomness.
 			return abTimers[random.Intn(len(abTimers))]()
 		}
 	}
